Introduce the means for issuing requests to the Alpaca API
This change introduces the Requestor struct. This struct abstracts away
from the lower level details concerning request submission and response
retrieval.
The Alpaca API we use is completely HTTP based. As such, we introduce
another abstraction which describes an individual endpoint and is used
by the Requestor to perform its job in a generic fashion.