Introduce support for interacting with Alpaca's /v1/account endpoint
As a first step on the long road towards automated trading against the
Alpaca API, this change introduces the means to query account
information as provided by this service. It does so by sending a GET
request to the /v1/account endpoint and reading and parsing the result
into the appropriate struct.
Among other things, the /v1/account endpoint provides information about
the user's purchasing power, which will be of importance once we start
trading.