Introduce supporting macro infrastructure
The previously introduced Trader struct for the alpaca module makes it
very easy to issue requests to the Alpaca service. However, in order to
do its job it needs an Endpoint definition it can work with. Defining
such an object is not quite trivial. Plus, we want to make the process
as declarative as possible, in order to allow for easy conversion from
an existing endpoint description as provided for many APIs into
something that our program understands.
To accomplish this goal (at least to a certain extent), we introduce a
macro that essentially accepts a description and does all the heavy
lifting of creating the proper types needed.