Add support for submitting an order
Now that we can access basic account data, the next step for us is to be
able to handle orders and everything surrounding them.
Hence, this change adds support for submitting an order. As part of the
submission an order ID is created and returned along with a
representation of the full order in the response.