Add support for canceling orders
This change adds support for canceling orders. Canceling works by
issuing a DELETE request to the /v1/orders/<order-id> endpoint, where
the order ID can be obtained from the order object retrieved in response
to the submission.
Along with that we add a tests for order submission (and cancellation,
of course), now that we have the means to clean up properly.