Use block_on_all over explicit runtime instantiation
As it turns out tokio already provides us with a shortcut for
instantiating a runtime and waiting for a request to finish via the
block_on_all function.
Compared to Runtime::block_on, which we have used so far, this function
has the advantage that it will wait for any spawned tasks as well, not
just the future passed in directly.
Given that property, we also change the way our task deletion in our
submission test cleanup works from an explicit issuing of Delete to the
spawn of a Delete task. Doing so has the advantage that in more complex
setups where similar cleanup is required it is less likely that we have
an error path where the cancellation does not happen.