Introduce RequestorExt helper trait
We anticipate that in the future we will need to issue more "fake"
orders for testing purposes.
In order to reduce code duplication and to make the code more concise
and to the point, this change introduces a testing-only helper trait,
RequestorExt, that extends the Requestor with a method to submit an
order.