Add support for retrieving a list of open orders
There is one last operation that we need to support in order to be able
to claim pretty much full support for orders (to the degree we may need
it) and that concerns the retrieval of all open orders.
This change adds the necessary definitions that allow us to do that. By
issuing a GET against /v1/orders we will get back a list of open orders.
We parse that into a vector of the already existing Order struct.