Make API base URL configurable
In the future we may want to trade with real funds, in which case we
need to communicate with a different API URL. Right now we also may want
to use this API when testing certain behavior.
In any case, this change adds support for configuring the API URL to use
through an environment variable.