Factor out env information retrieval
In the past we have decided to keep mechanisms consistent with that of
Alpaca SDKs and libraries and use environment information for passing
certain data and secrets into the program.
In the future we will need to instantiate another entity that needs to
authenticate with the Alpaca API, in addition to the Requestor. Hence,
with this change we factor out the environment related functionality
into a separate module, such that it can be reused more easily.