Correctly check for errors on spawned tasks
As it turns out we were missing an error check on the async order
cancellation path we use in the tests. In particular, we spawn the
deletion task but never evaluate its success.
With this change we correct this issue. Along with that we move this
functionality into the RequestorExt trait to allow for easier sharing.