Log potentially relevant debug information
We want to be sure that when used in production we have enough
information available to debug problems thoroughly. As such, it is
useful to have the possibility of just incrementing a log level and
having more information appear.
In order to accomplish that, this change adds a dependency to the log
crate and adjusts the Requestor to log potentially relevant information.

Furthermore, because in many cases logs are equally useful for testing
in order to understand what is going on, by using the test-log-env crate
this change adjusts all tests to initialize env_logger to allow for easy
accessibility of logs.