Rework Endpoint trait
The interface between the Requestor and the Endpoint is somewhat
improvised, with the Requestor calling into the Endpoint at a variety of
points during the formation of the final request. Ultimately, however,
it should not be the Requestor that creates the request, but the
Endpoint itself. Doing so leads to much less back and forth and a
cleaner and arguably more descriptive interface.
The main reason for the current design is that the Requestor holds the
authentication information and injects it into the request being built.
However, ultimately that should not be its job. Conceptually, while it
is the entity with this information, it should just make it available to
whomever needs it to create the final request.
With that in mind this change refactors Endpoint trait. Most
importantly, the Requestor now only calls into Endpoint::request to
create the actual request, and this method accepts the API base URL as
well as the authentication information along with the actual input. The
remaining methods are mostly used to implement Endpoint::request in a
sensible fashion by default. Along with this change we remove the
Endpoint::builder method, which no longer serves a purpose, and add two
new ones to describe the request to be created: Endpoint::method and
Endpoint::body.