Consolidate endpoint related functionality into endpoint.rs
In most cases it is preferable to organize modules by what they contain
(i.e., which part of the program they pertain to) rather than how they
accomplish that.
The macros.rs file does not adhere to this ideal. Hence, with this
change we rename macros.rs to endpoint.rs (because it really contains
functionality pertaining the handling of endpoints). Furthermore, we
move the definition of the Endpoint trait into the newly created
endpoint.rs file.