Rename Requestor to Client
The name Requestor has always been a fairly specialized and
unconventional term, given the kind of operations the struct performs.
Not only that, similar but not really request related functionality will
be added to the struct, making the name no longer well suited for the
final result.
A more commonly used and more generic name for the functionality it
provides is the term client. Hence, with this change we rename the
struct to Client.