Add data structures used during websocket stream handshake
Having made available the core trading related request and response
primitives of the Alpaca API, the next step is to subscribe to the
event streams provided.
As a first step we are interested in two such streams: one for account
updates and another for trade updates.
The way we subscribe to these streams is by going through a specific
handshake that
  1) authorizes the streaming and
  2) actually subscribes to specific event streams

This change introduces the data structures used for representing the
data exchanged during this handshake. These data structures will handle
all the serialization and deserialization for us.