Implement websocket event stream subscription handshake
Now that we have all the data structures available to properly encode
handshake requests and decode the corresponding responses, this change
adds support for actually subscribing to such a event stream.
In particular, it introduces the first version of a function that will
yield a futures-style Stream representing the individual decoded events.
Right now, though, it just yields websocket crate's OwnedMessage.