Actually decode and stream the proper event objects
With the streaming scaffolding in place and the handshake already
happening properly, this change now adds the logic to properly decode
the individual websocket messages into the desired event objects. In
order to make that happen while correctly handling messages that
indicate a close of the websocket connection (which should terminate the
stream) we have to introduce an indirection into the otherwise fairly
straight forward logic.