Finish end-to-end support for streaming account and trade updates
We now have all the low level plumbing in place to finally define the
actual event objects that we will receive. Each of the two channels
(for account updates and trade updates) has their own event type.

With this change we define those event types, plumb the lower level
streaming functionality into the Client which provides a concise
interface for accessing it, and add an end-to-end tests that subscribes
to the trade update stream on Alpaca's paper trading API, submits an
order, and checks for the expected event to occur.