Add mock server and tests
Despite being superficially ready, the streaming functionality still has
a few rough edges and cases that are not well covered. To identify those
and to prevent regressions due to future changes, this change introduces
a mock websocket server that can be used to trigger certain error cases
in the streaming code. Along with that various test cases are added that
do so already.