Properly respond to Ping messages
The initial version of the streaming code did intentionally not support
responding to Ping messages as they are part of the websocket protocol.
This change makes good for this sloppiness by correctly responding back
to such messages with a Pong message.
The difficulty here is that we cannot (or rather: do not want to) simply
"spawn" a new background task as is usual in the future ecosystem. Such
a spawn is specific to the reactor or executor and we do not want to tie
this crate to any specific implementation (think, Tokio). Hence, we send
the Pong as part of an intermediate future whose result we pretty much
ignore.