Do not abort stream on anything but websocket errors
As it turns out, reporting an error on the stream actually does
stop the stream -- in the sense that a for_each will stop, for example.
Now, it may be possible to start polling it again (conceptually, the
futures::Stream trait apparently allows for that, but it appears that
futures::unfold does not), but that really is a cumbersome workflow. We
are mostly fine with a websocket related error terminating the stream in
this way (the assumption here is that there really should not be any
transient errors, but that may be a false assumption), but a JSON decode
error should be no means lead to such a cancellation.
Hence, with this change we rework the event stream logic to actually
yield Result objects instead of items. This way we can package up
decoding errors inside the future's item and push handling those to
clients, which arguably is where such functionality should reside.