Simplify definition of endpoints
In order to define an HTTP endpoint in a declarative manner we have
introduced the EndpointDef macro. This macro expects a total of three
types: the result being returned by the endpoint, a wrapper for this
result, and an error type.
As it turns out the wrapper type is not actually needed (anymore?). It
is not quite clear why we would ever have needed this type, but the
believe is that this had to do with trait implementations not being
possible. In any case, since this problem no longer actually exists,
this change removes this intermediary, which just resulted in additional
complexity and potentially confusion.