Introduce endpoint error variant for 401 Unauthorized responses
As it turns out the undocumented 401 Unauthorized HTTP status can be
returned on what are believed to be all endpoint requests (as all of
them have to carry authentication information). Right now such responses
would just be classified as an unexpected HTTP status but we want such a
result to be a first class entity.
To that end, this change adjusts the EndpointDef macro to automatically
generate such an error variant.