Narrow down Endpoint related errors
The Endpoint trait is supposed to be as generic as it gets, to the point
that it could easily be moved out of the crate. One thing that currently
stands in the way of doing that is its reliance on the crate's Error
enum.
With this change we introduce a new error type, EndpointError, that
comprises the actual variants that the trait deals with. This step both
ensures better decoupling while also making it clearer what error
variants a caller can expect to be emitted.