Decode and store Alpaca reported errors
When issuing a request to Alpaca that results in some form of error, the
service actually responds back with an error code and a message in the
body that describe the problem encountered more closely. So far we have
just ignored this data, but it can actually be very useful.
To that end, this change takes the necessary steps to parse and store
this message so that the user can evaluate it.