Implement std::error::Error for all error types
With the recent change to add an additional Alpaca provided error
message to each endpoint related error variant, formatting those
properly has become cumbersome and nothing that we want to push to all
clients.
To remove this burden from clients this change provides a reasonable
Display implementation for all such errors. Also, now that we implement
Display we may as well implement std::error::Error, which provides
clients with the option of treating errors as trait objects.