Ensure proper documentation of endpoint error variants
The EndpointDef macro we use for declaring an endpoint along with its
expected errors currently lacked support for adding proper doc strings
to each of those errors.
With this change we fix this deficiency. We adjust the macro to accept
any such meta items and remove the allowance of missing documentation
for the generated error struct.