Properly parse response time stamps
A couple of objects returned by the Alpaca API contain time stamps. So
far we have shied from actually parsing them, mostly because that
involves yet another dependency to rely upon.
Yet, with this change we add code to properly deserialize them, instead
of just storing them as strings. We use the chrono crate in order to
make sense of the values, but we ultimately store them as
std::time::SystemTime objects, mostly for two reasons:
1) to keep this dependency internal and not user-facing and
2) because the hope is that comparison of time stamps is enough, which
   is provided by SystemTime objects already