Encapsulate Alpaca API related information in ApiInfo struct
This change introduces a new struct that encapsulates and groups
together the three bits of information (the base URL, the key ID, and
the secret) needed when communicating with the Alpaca API: ApiInfo. The
main purpose of this struct is to allow for dependency injection as well
as to decouple retrieval and passing around of this information from
entities using it.
Previously, the Client, for example, was more or less directly
retrieving this information from the environment, whereas now it just
stores and uses an ApiInfo object -- where that object came from does
not matter.
Another driver for this change is the need to be able to pass around
(and potentially use) this information separate from the Client itself.
The Client struct contains additional data, which hinders it from
implementing Clone. The ApiInfo does not have such a limitation.