Expose WebSocketError directly
In the past we have converted certain errors reported by the websocket
crate into our own errors. The down side to this approach is that it is
less obvious what errors a particular function can actually emit.
Depending on the complexity of the function, that may or may not be a
concern.
For the streaming functionality it is difficult to have a good
understanding of the different errors at the various layers and so
direct exposure of the error in question can ease understanding. Hence,
with this change we expose those websocket errors directly, instead of
embedding them into our own Error type.