Publicly expose low-level stream function
This change exposes the "low-level" stream function that is used for
streaming account and trade updates from Alpaca. While this function is
indirectly exposed through the Client interface already, that may not
accommodate all use cases: The Client contains a hyper::Client, which
cannot be cloned. However, in future-style programs it may be necessary
to subscribe to a stream when the original client has already been moved
into a different part of the program.
Using the recently introduced ApiInfo object, which is freely clonable,
in conjunction with the stream function we are able to stream such
updates independent of the Client. To not require users to recreate a
suitable ApiInfo object, we also adjust the Client to allow access to
its version.