Introduce builder for creating custom Client objects
In the past we have had to work around the fact that the hyper crate
by default keeps around idle connections which in turn are managed
internally in such a way that a block_on_all call will actually wait for
them to die.
This was unacceptable for tests as it would make them appear as hung. So
we explicitly changed the defaults for tests in such a way as to no
longer keep idle connections around.
However, this problem is not really restricted to tests in this crate,
but can also be witnessed in other crates using this one. To allow users
to fix this problem there too, we introduce a Builder for Client objects
that allows for easy configuration of this option. Note that to keep our
own tests concise we still keep the compile time adjustment of the
Builder's defaults locally.