Introduce a dedicate type representing an asset's symbol
Alpaca supports different representations for an asset's symbol. In one
instance, it can really just be the symbol as a string, but there are
also forms that include the exchange and assets class. So far we have
not differentiated between those variants and left it up to the user to
decide which one to use and how to format it.
With this change we introduce a dedicated enum type, Symbol, that
ensures proper representation at the type level. This new type is only
used for requests, though, because Alpaca's responses always provide the
various fields (symbol, exchange, asset class, and asset ID) separately.