Track account status through Status enum
This change introduces a new enum that represents the account's status.
So far we have merely passed through the status reported through the
Alpaca API in the form of a string. That was sufficient because we did
not actually evaluate it anywhere. However, because supporting this
additional field is low effort we finally do so.