Add support for querying the /v2/account endpoint
Alpaca recently released v2 of their API, with the main new features
being margin and premarket/after hours trading.
As a first step towards supporting this version of the API, this change
introduces the means to interacting with the /v2/account endpoint.
Objects returned are very similar to v1 ones, but contain a few
additional fields and deprecated one, portfolio_value, which is removed
from our definition entirely.