Convert stream tests to using an async WebSocket server
We currently use a synchronous style mock WebSocket server for testing
our implementation of a stream. That is a bit clunky and does not play
well with our desire to use async/await syntax in the future.
To that end, this change convert the mock server to use the asynchronous
server functionality provided by the websocket crate instead.