Rename tokio dependency to tokio01
Our testing infrastructure currently uses the tokio crate for driving
futures to completion. With recent changes surrounding async/await in
the Rust ecosystem there are now two major versions of this crate
around: 0.1 uses legacy futures and 0.2 uses std::futures::Future. While
crates are generally moving towards using the latter, not all are there
yet.
With upcoming changes we want to catch up and add support for the
async/await syntax recently stabilized. This will require us to use
std::futures::Future and, hence, tokio in version 0.2. Until all crates
we use are converted to using this version of tokio we need to support
0.1 as well.
To that end, this change renames the tokio dependency we consume to
tokio01, in anticipation of tokio 0.2 being referenced merely as tokio.