Rename futures dependency to futures01
Very similar to tokio (and for similar reasons), this change renames the
futures dependency we use to futures01. Note that while
std::futures::Future is stabilized, we will require the futures crate
compatibility reasons (bridging between the two worlds). Renaming the
crate also makes it obvious that it is meant to be removed as soon as
possible.
While at it, this patch also disables the default features of the
futures crate.