Replace ClientExt trait with free standing functions
This change removes the ClientExt trait and makes its methods,
order_aapl and cancel_order, free standing functions. This step is taken
because of limitations surrounding trait impl return types in traits: we
have to box up the future and with a future change there are some
pin/unpin issues that we could not figure out.