Migrate (almost) fully yo async/await
This change moves the majority of the crate to async/await syntax.
Unfortunately it is rather huge, because of all the (non-obvious)
interdependencies among components. While it would be possible to have
meaningful smaller intermediate steps, that requires a lot of work to do
properly.
Perhaps the most tricky part is the tokio runtime. tokio 0.1 and 0.2 are
not compatible. In order to get the benefit of the async/await syntax we
have to use the #[test] attribute, which is specific to tokio 0.2. The
version of hyper we use, however, requires tokio 0.1. And in fact, it
spawns futures/tasks using the tokio 0.1 runtime and that will fail in
functions with tokio 0.2 style #[test].
While there exists a tokio-compat crate, it does not come with a #[test]
attribute. What's worse, async closures, which would be an alternative
syntax, are not yet stable.
Because of all that messiness, this change also updates hyper to version
0.13, which now uses tokio 0.2.

In doing so we now have async Client::issue and Client::subscribe. This
gives us the benefit of being able to use the await syntax, which makes
the code become a little less complex. One side effect of that
simplification is that we removed the cancel_order utility function.

Unfortunately the streaming and request issuing API tests have a bit of
overlap in terms of functionality they require, which is the reason for
why those two components could not be converted in isolation either.
Moreover, while converting from a future in version 0.1 to 0.3 is
straightforward using the futures-compat crate, the reverse is not
necessarily true (there are some strange unpin issues that we failed to
resolve). If that were straight forward, a bit more of the streaming
functionality could have been converted in advance.