Convert part of the streaming functionality to use Stream 0.3
In an attempt to cut down on the number of usages of futures 0.1 (and in
order to go with the time), this change converts part of the streaming
functionality we have from using the Stream trait in version 0.1 to that
in version 0.3.
Note that the majority of uses of this trait are to be found in the
WebSocket functionality we use, and that cannot change until upstream
has taken the necessary steps as well.