Remove v1 account support
This change marks the first step towards fully moving towards v2 of the
Alpaca API. Upstream has deprecated the v1 API and we intend to remove
support for it, as v2 is working as expected by now.
To that end, this starts off the process by removing the v1::account
module, integrating required functionality directly into v2::account.