Implement FromStr for asset::Symbol, asset::Class, and asset::Exchange
Clients using this crate may encounter cases where a symbol representing
an asset is supplied in the form of a string and it should be parsed
into the appropriate object. It makes arguably little sense to have
each and every client re-implement this logic by itself.
Hence, with this change we provide an implementation of FromStr for the
asset::Symbol enum (as well as asset::Class and asset::Exchange). The
asset itself can be specified in the various forms supported by the
Alpaca API:
- symbol
- symbol, exchange
- symbol, exchange, class
- ID