Migrate mock_server over to using tungstenite websockets
As it turns out the websocket crate we have been using for so long has
not kept up with changes surrounding async/await and, in fact, has a
notice saying that "maintainership is sluggish". It's not clear what
exactly this means, but it does not look great.
In the future we will need more WebSocket code for tick streaming and it
makes sense to use the same crate in all cases. At the same time, it
seems wrong to decide in favor of the websocket crate for this purpose,
given the above notice.
The only real alternative at this point seems to be tungstenite. It is
fundamentally synchronous, but there exist asynchronous "bindings".
Those in turn come in two flavors: one for the async-std eco system and
another for tokio. The fact that both exist is in and of itself a huge
problem, because it means that there is a split in the community. Most
relevant for us, though, is the fact that tokio-tungstenite also lacks
behind and only supports futures 0.1.
Hence, the only logical decision at this point is to use
async-tungstenite. This change does exactly that and rewrites the
websocket streaming server used for testing purposes using this crate.
Doing so we can finally drop tokio 0.1 as a dependency (although now we
also depend on asynd-std, which is an even bigger dependency magnet from
what it seems.