Migrate WebSocket client code to tungstenite
This change migrates the WebSocket client code that we have for
streaming account and trade events from using the websocket-rs crate to
tungstenite.
It is necessary because because websocket-rs is seemingly no longer
maintained properly and it has not yet caught up to supporting
async/await syntax (which we want to use). As it stands it appears that
tungstenite is the only viable alternative implementation of the
WebSocket protocol. With websocket-rs no longer being used, we also
dropped the last consumer of futures 0.1.
Note that this change is a quite literal conversion, meaning that
existing constructs were converted with as little churn as possible.
Given that everything is using async/await now, a subsequent change will
attempt to make the code nicer in conjunction with that.

Aside from futures 0.1 code being horrible to maintain (and that was
very much visible in the work leading up to this change), there was one
severe complication that is probably worth pointing out. Futures
returned by async functions and async blocks for some reason seemingly
do not implement Unpin. That is a problem for certain functionality
which does require this marker trait. Perhaps most importantly the next
method on a Stream (or the into_future alternative) both require Unpin
and in fact one of our tests did hit this problem.
Now conceptually it would be possible to not use async functions
"inside" the stream (meaning that we would fall back to using
combinators everywhere; hello futures 0.1!), which presumably would keep
it from not implementing Unpin. However, that was not possible either
because of the brain dead change of the signature of the Sink::send
method, which no longer consumes the object (but works with a reference
to self). The result of that mess being that we cannot "inject" the
sink object back into the future (as we do for sending pongs in response
to pings) due to borrow conflicts. We also cannot spawn a task doing so,
because that in turn requires a future that is 'static.
Ultimately, we just bit the bullet and kept using async/await but lived
with a Stream not being Unpin.