Move authentication and subscription functionality into handshake module
With the migration towards using the tungstenite crate we noticed that
the separation between the stream module and handshake is not as minimal
as it could be.
To improve on the situation a bit, this change moves more of the
authentication and subscription logic into the handshake module,
allowing the stream module to focus on the main stream handling related
logic.