Prepare streaming functionality for refactoring
Because we require similar message/event streaming functionality on top
of a WebSocket channel in a different crate, we need to decouple that
functionality from the specifics of this crate.
Lucky for us, the code is in a pretty good shape already. The only part
specific to this crate is usage of our internal stream::Event struct,
which we use as a template for the JSON decoding. We will still need
this part moving forward, but with this change we move all references to
this type further out into the non-generic parts of the stream creation.