Store key ID and secret as strings
When working textual data that is coming in as raw bytes, there is
always a tension in Rust between converting early, late, or perhaps not
at all.
In the particular case at hand we get data from the environment and
there is no obvious reason to ever convert it into a string. Except that
deep down we need the textual representation in order to tell the JSON
encoder to not encode the data as an array of bytes. Because of that we
were having an unwrap in the authentication path.
In general it is better to catch potential errors earlier, and so with
this change we keep the data around as a true string (as opposed to a
vector of bytes we had beforehand) and convert into that as the first
thing after reading the content from an environment variable.