Simplify event streaming functionality
This change simplifies the event streaming functionality a bit by moving
the URL modifications into the client and removing the testing-only
stream_insecure function.
With the switch to using the tungstenite crate it is the URL's scheme
that determines whether to use TLS or not, not the function we invoke to
establish the connection. As such, we can just unconditionally use a TLS
connector and tungstenite will decide whether to actually use an
encrypted connection.