Automatically implement Endpoint trait in EndpointDef
The functionality for defining an endpoint in a declarative manner is
something that we want to make generally available to facilitate reuse.
Thinking about how that is possible, it appeared that it would be better
to not have the user explicitly define an implementation of the Endpoint
trait for the type at hand. Rather, the EndpointDef macro we have can
accomplish this task. This change does exactly that, it adjusts the
macro to automatically implement the trait, reducing the amount of code
that the user has to write.