Move request creation functionality out of Endpoint trait
The task of making the Endpoint trait more generally usable is hindered
by two more road blocks:
1) The ApiInfo struct we currently require for the creation of a
   request is specific to the crate the endpoint is defined in.
2) The notion of adding authentication information to a request, which
   can but does not have to be based on HTTP headers, is not yet
   general enough.

This change sort of side steps these problems by essentially moving the
request creation functionality out of the Endpoint trait altogether.
Instead it is up to the client to create a request object based on the
information provided by an implementation of the Endpoint trait. As
such, the Endpoint trait will no longer have any "active" component to
it, but solely be used for describing properties of an endpoint. Another
consequence of this decision is that we can nuke the EndpointError
error type.

As a bonus, this change removes an unwrap from the client code. To do so
we need to equip errors defined for an endpoint with an additional
variant containing "regular" HTTP errors (i.e., not hyper ones).