Consolidate associated type constraints in Endpoint enum
There are certain constraints that we impose on the associated types
contained in the Endpoint trait. Some of those are more or less implicit
in the macro generated code, while others are stated explicitly in the
client code. Given that those constraints are ultimately shared amongst
all users of this crate, it makes sense to explicitly state them in the
trait definition directly, eliminating the need to do so on individual
"call" sites.