Introduce Endpoint::parse_err method
This change introduces a new method to the Endpoint trait: parse_err.
This method is used to parse an API related error from an HTTP body.
The explicit definition of this method is mainly done to mirror what we
have with the preexisting parse method (which is what we use in the
happy case). Doing so also pushes all JSON parsing code into the trait
itself, as opposed to having half of it in the generated code.