Implement Display for ErrorMessage
The EndpointDef macro generated code still depends on the ErrorMessage
type. While that is probably not something that we can change moving
forward, we can at least make the coupling between the two less tight.
This change does so by no longer generating an ad-hoc inline definition
of how to format such an error, but instead relying on a proper
implementation of Display.