Introduce ApiError type to Endpoint trait
This change introduces another associated type to the Endpoint trait:
ApiError. This type is meant to represent errors that are reported by
the API itself, on expected and unexpected HTTP status codes alike.
So far the ErrorMessage type implicitly fulfilled this role, but with
this change we make the relationship explicit. Definition of this type
is a bit tricky, as we do not want to have it appear in every endpoint
definition. However, we can fudged this case with a mechanism we already
employ and intend to keep: the wrapping of the macro and overwriting of
chosen parts of it. More concretely, in the current design the
EndpointDef macro would be a crate specific macro definition that could
set this type to the crate's error type. It is only the EndpointDefImpl
macro that is fully generic.