Properly implement std::error::Error::source for custom errors
The way we implement Display for our custom error types does not seem
truly correct, at least not for a library crate. We print the error
along with all sources. However, users of the crate may reasonably
expect to be able to iterate over the sources themselves and evaluate
them. So far this has not been supported, because we stuck to the
default implementation of std::error::Error, which returns no sources
for any given error.
This change reworks the state in such a way that we no longer print all
sources as part of our implementation of Display. At the same time we
allow clients to do so because we implement proper forwarding of source
information to the caller.
In an application crate the original behavior is arguably easier to deal
with, as there is no explicit need to iterate over the sources. But
given that we are working on a library crate, we do not want to make
such a policy decision.