Remove ConvertResult intermediary type
In the past we were required to introduce an intermediate type in order
to convert the result of a request, in the form of a (status, body)
pair, into either an output (if the status indicates a successful
operation) or an error (otherwise). This has never been a pearl of API
design as it is hard to understand and just feels clunky.
Now that we have transitioned towards a scheme where the macro we define
is used also for implementing the Endpoint trait, we can rectify this
shortcoming. With this change we move this conversion logic into a
method in the trait. This method, contrary to all others, should not be
overwritten, as it is automatically generated unconditionally.