Automatically define request struct in EndpointDef macro
The endpoint functionality rises and falls with the actual struct
representing the request type. So far this type definition had to be
created explicitly by the user. This patch changes this fact by
adjusting the macro to emit the struct definition itself. Doing so has
the advantage that users no longer need to write this code (which is
repetitive and essentially just templating).