Accept visibility specifier in EndpointDef macro
Now that the EndpointDef macro also emits the struct definition we need
a way to influence visibility of the struct to make the macro usable
from a more general context as well.
To that end, this change adjust the macro to accept a visibility
specifier and adjusts all endpoint definitions to declare their endpoint
type public (as they implicitly were beforehand).