Use http-endpoint crate for endpoint definition
With this change we migrate all of the non-Alpaca specific endpoint
related functionality into a new crate, http-endpoint, and use that
crate from this one.
Unfortunately there is still one (non-showstopper) problem that we have
been unable to address, and that is about error conversions. Similar to
before, we create a dedicated error type for each endpoint definition.
Unfortunately, though, converting those into our own crate's type is no
longer as straight forward, making the question mark operator more
cumbersome to use. Right now we added an explicit 'from' conversion, but
it is not nice. Perhaps in the future we find a better way.