Add support for short positions
As it turns out Alpaca does not explicitly distinguish between the
intention to short an equity or to go long on it. If one sells an equity
that is not currently owned, a short position will be created.
In order to fully support short positions, however, we also need to
adjust the 'Side' enum, which is used in the response to a request
listing all open positions. This change takes care of that.