Use tracing crate for providing tracing functionality
For a long time now we had the plan to provide better introspection
capabilities in the form of tracing. Tracing captures additional context
that is vital for understanding emitted messages in an async program. We
have finally come around to investigating possible options on how to go
about that.
There are two main tracing crates, tracing and async-log.

Async-log, seemingly coming from the asynd-std guys, doesn't bring much
value to the table to begin with. The good thing is that it uses the log
crate and so our existing infrastructure surrounding it would "just
work". However, we have seen a pull request adding async-std as a
dependency, making this an extremely dangerous choice. That, combined
with a questionable value-add, disqualify it.
The tracing crate is fairly sophisticated. So sophisticated, in fact,
that it took a day or so to understand how to work with it and convince
an application to log traces. There are various ways to enable logging:
- the "log" feature enables rudimentary logging of all events/spans;
  this logging is so rudimentary, in fact, that it isn't even useful:
  context is not included properly and so the value-add over mere
  logging is close to zero
- the tracing-log crate comes with some support for env_logger-like
  initialization -- but that didn't really work as intended: no messages
  appeared
- tracing-subscriber comes with a nicely configurable formatter that
  emits logs (with span contexts) to stdout; it can be configured
  env_logger-style

From the above three possibilities only tracing-subscriber is a viable
option. The only alternative is to roll our own tracing in a more ad-hoc
manner. The problem with that approach is that it will be hard the
moment we want to trace across crate boundaries.
As such, this change uses the tracing crate. Given that we are working
in the context of a library crate we do not care much how the traces are
consumed, but just how they are emitted. Except for tests, which now
require test-env-log in version 0.2.2 (that got support for initializing
tracing using the tracing-subscriber crate).
Note that some refactorings had to take place in order to make
everything work as expected.